test: reorganize tests/ into per-source-layer subfolders (Phase 4) (#254)
Mirror ams/ source layout in tests/. Each subfolder holds tests
for one slice of the source tree:

  tests/system/    ams/system.py          (System aggregate, addressing, repr, warnings, paths, param correction)
  tests/io/        ams/io + outputs       (load, case catalogue, export csv/json, report)
  tests/core/      ams/core/              (model, routine_ns, matprocessor, service)
  tests/models/    ams/models/            (group registry)
  tests/opt/       ams/opt/               (omodel — CVXPY wrapper layer ONLY)
  tests/routines/  ams/routines/          (Routine base + variants + parametrized scenarios)
  tests/interop/   ams/interface.py       (andes bridge, with TestJumperToAndes folded in)
  tests/cli/       ams/cli.py             (CLI surface only — argv, version, dispatch)
  tests/bench/     ams/bench/             (benchmark suite smoke tests)
  tests/codegen/   docs/codegen           (joins existing test_generator.py)

Three normalizations done in the same hop:
  test_1st_system.py  -> split: TestCodegen → codegen/test_first_run_docs.py
                              TestParamCorrection → system/test_param_correction.py
  test_interface.py   -> interop/test_andes_interface.py
  test_matp.py        -> core/test_matprocessor.py  (was wrongly named "matpower" — subject is MatProcessor)
  test_routine.py     -> routines/test_routine_base.py  (clarifies it's the Routine base API, not a single routine)
  test_jumper.py      -> folded into interop/test_andes_interface.py as TestJumperToAndes (also fixes pre-existing TessJumper class-name typo)

Principle: tests/<layer>/ tests subject under test from ams/<layer>/.
tests/cli/ stays focused on CLI surface; details behind CLI live in
the dedicated subject test (test_report → io/, test_bench → bench/).
tests/opt/ holds CVXPY-wrapper tests only; non-CVXPY ams/core/
machinery (matprocessor, service, model, routine_ns) lives in
tests/core/.

No test bodies, fixtures, or imports change. tests/conftest.py stays
at the top so per-subdir tests inherit the same shared fixtures
(case loaders, benchmark_mpres, HAS_MISOCP/HAS_PYPOWER flags).

Suite still 348 passed. flake8 clean.

CI workflows invoke `pytest` with no path arg, so subdir layout is
auto-discovered without any workflow change.
