+++ tests/core/test_matprocessor.py
