+++ tests/core/test_routine_ns.py
