+++ tests/core/test_service.py
