+++ tests/io/test_case.py
